done with basic learning

+++ run.sh
@@ -17,25 +17,33 @@ else
 
                         dirpath=$(dirname "$filepath")
                         basename=$(basename "$filepath")
+
                         without_ext="${basename%.*}"
 
                         echo "[√] Compiling object file..."
                         cd "$dirpath"
-                        nasm -f elf32 -o "$without_ext.o" "$basename"
 
-                        if [[ $? -eq 0 ]]; then
+                        lastdir="${dirpath##*/}"
+                        if [[ "$lastdir" != "baremetal" ]]; then
+                                
+                                nasm -f elf32 "$basename" -o "$without_ext.o"
+                                if [[ $? -eq 0 ]]; then
 
-                                # ld -m elf_i386 -s -o "$without_ext" "$without_ext.o"
-                                gcc -m32 "$without_ext.o" -o "$without_ext" 2> /dev/null
-                                echo "[√] Done"
+                                        # ld -m elf_i386 -s -o "$without_ext" "$without_ext.o"
+                                        gcc -m32 "$without_ext.o" -o "$without_ext" 2> /dev/null
+                                        echo "[√] Done"
 
-                                "./$without_ext"
-                                rm "$without_ext.o" "$without_ext"
-                                cd - > /dev/null
+                                        "./$without_ext"
+                                        rm "$without_ext.o" "$without_ext"
+                                        cd - > /dev/null
 
+                                else
+                                        echo "[x] Compiling failed"
+                                        exit 1
+                                fi
+                        
                         else
-                                echo "[x] Compiling failed"
-                                exit 1
+                                nasm -f bin "$basename" -o "$without_ext.bin"
                         fi
 
                 else
